# swagger.yaml
...
openapi: '3.0.0'
info:
  title: Aberrations Sheets API
  description: API for Aberrations Sheets application.
host: api.sheets.aberrations-rpg.com
basePath: /v1
paths:
  /players/:sheetType:
    get:
      summary: List all of a players character or campaign sheets (unaggregated).
      tags:
        - Players
        - Character Sheets
        - Campaign Sheets
      parameters:
        - in: path
          name: sheetType
          required: true
          schema:
            $ref: '#/components/schemas/sheetType'
    post:
      summary: Create a character or campaign sheet for a player.
      tags:
        - Players
        - Character Sheets
        - Campaign Sheets
      consumes: application/json
      parameters:
        - in: path
          name: sheetType
          required: true
          schema:
            $ref: '#/components/schemas/sheetType'
        - in: body
          name: sheet
          description: The character or campaign sheet to create.
          schema:
            oneOf:
              - $ref: '#/components/schemas/CharSheet'
              - $ref: '#/components/schemas/CampSheet'
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
  schemas:
    CharSheet:
      type: object
      required:
        - playerName
        - characterName
        - charBackground
        - charDescription
      properties:
        playerName:
          type: string
        characterName:
          type: string
        charBackground:
          type: string
        charDescription:
          type: string
        lifeKredits:
          type: number
    CampSheet:
      type: object
      required:
        - name
        - overview
        - details
      properties:
        name:
          type: string
        overview:
          type: string
        details:
          type: string
...
    sheetType:
      type: string
      enum:
        - characters
        - campaigns

# Apply the security globally to all operations
security:
  - bearerToken: []